requirements.txt
streamlit==1.36.0
langchain==0.2.5
langchain-community==0.2.5
langchain-openai==0.1.8
openai==1.35.7
python-dotenv==1.0.0
streamlit-authenticator==0.3.2
pytest==7.4.3
PyYAML==6.0.1 